[#13648] - Reformatted tests
