+++ tests/unit/Firewall/Adapter/Acl/SetCacheCest.php
@@ -12,20 +12,176 @@ declare(strict_types=1);
 
 namespace Phalcon\Test\Unit\Firewall\Adapter\Acl;
 
+use Codeception\Example;
+use function ob_end_clean;
+use function ob_start;
+use Phalcon\Acl as PhAcl;
+use Phalcon\Acl\Adapter\Memory;
+use Phalcon\Cache\AdapterFactory;
+use Phalcon\Events\Manager;
+use Phalcon\Firewall\Adapter\Acl;
+use Phalcon\Mvc\Dispatcher;
+use Phalcon\Mvc\Model\Binder;
+use Phalcon\Storage\SerializerFactory;
+use Phalcon\Test\Fixtures\Firewall\BindingRole;
+use Phalcon\Test\Fixtures\Traits\DiTrait;
+use Phalcon\Test\Fixtures\Traits\FirewallTrait;
+use Phalcon\Test\Models\AlbumORama\Albums;
 use UnitTester;
 
 class SetCacheCest
 {
+    use DiTrait;
+    use FirewallTrait;
+
+    /**
+     * @var Dispatcher
+     */
+    protected $dispatcher;
+
+    /**
+     * @var Acl
+     */
+    protected $firewall;
+
+    public function _before()
+    {
+        ob_start();
+
+        $this->setNewFactoryDefault();
+        $this->setDiMysql();
+
+        $dispatcher = new Dispatcher();
+        $dispatcher->setDefaultNamespace(
+            'Phalcon\Test\Controllers\Firewall'
+        );
+        $dispatcher->setDI($this->container);
+        $eventsManager = new Manager();
+        $eventsManager->attach(
+            'firewall:beforeException',
+            function () {
+                return false;
+            }
+        );
+        $firewall = new Acl('acl');
+        $firewall
+            ->setEventsManager($eventsManager)
+            ->setRoleCallback(
+                function ($di) {
+                    return $di->get('myrole');
+                }
+            )
+            ->setAlwaysResolvingRole(true)
+        ;
+        $dispatcher->setEventsManager($eventsManager);
+        $this->dispatcher = $dispatcher;
+        $this->firewall   = $firewall;
+    }
+
+    public function _after()
+    {
+        ob_end_clean();
+    }
+
     /**
      * Tests Phalcon\Firewall\Adapter\Acl :: setCache()
+     *
+     * @dataProvider getCache
      *
      * @author Phalcon Team <team@phalconphp.com>
      * @since  2019-05-12
      */
-    public function firewallAdapterAclSetCache(UnitTester $I)
+    public function firewallAdapterAclSetCache(UnitTester $I, Example $example)
     {
         $I->wantToTest('Firewall\Adapter\Acl - setCache()');
 
-        $I->skipTest('Need implementation');
+        $acl = new Memory();
+        $acl->setDefaultAction(PhAcl::DENY);
+        $acl->addComponent("Four", ['first', 'second']);
+        $acl->addComponent("Three", ['deny',]);
+        $acl->addRole('ROLE1');
+        $acl->addRole('ROLE2');
+        $acl->addRole('ROLE3');
+        $acl->addRole('ROLE4');
+        $acl->addRole('ROLE5');
+        $acl->allow('ROLE1', 'Four', 'first');
+        $acl->allow('ROLE1', 'Four', 'second');
+        $acl->allow('ROLE1', 'Three', '*');
+        $acl->allow('ROLE2', 'Four', 'second');
+        $acl->allow(
+            'ROLE2',
+            'Four',
+            'first',
+            function (BindingRole $user, Albums $album) {
+                return $user->getId() == $album->artists_id;
+            }
+        );
+        $acl->allow('ROLE3', '*', '*');
+        $acl->allow(
+            'ROLE4',
+            'Four',
+            '*',
+            function (BindingRole $user, Albums $album) {
+                return $user->getId() == $album->artists_id;
+            }
+        );
+        $acl->allow(
+            'ROLE5',
+            '*',
+            '*',
+            function (BindingRole $user) {
+                return $user->getId() == 3;
+            }
+        );
+
+        $binder = new Binder();
+        $this->dispatcher->setModelBinder($binder);
+        $this->container->set('acl', $acl);
+        $eventsManager = new Manager();
+
+        $serializer = new SerializerFactory();
+        $factory    = new AdapterFactory($serializer);
+        $cache      = $factory->newInstance('memory');
+
+        $this->firewall->setCache($cache);
+        $this->firewall->setRoleCacheCallback(
+            function (BindingRole $user) {
+                return $user->getId();
+            }
+        );
+        $eventsManager->attach('dispatch:afterBinding', $this->firewall);
+        $this->dispatcher->setEventsManager($eventsManager);
+        $dispatcher = $this->dispatcher;
+        $this->container->set('dispatcher', $dispatcher);
+
+        $returnedValue = $this->getReturnedValueFor(
+            $this->container,
+            $dispatcher,
+            $example[0],
+            $example[1],
+            $example[2],
+            $example[3]
+        );
+        $I->assertEquals($returnedValue, $example[4]);
+        $I->assertEquals($cache->get("_PHF_")[$example[5]], $example[6]);
+    }
+
+    private function getCache(): array
+    {
+        return [
+            ["four", "first", new BindingRole("ROLE1", 1), ['album' => 1], "allowed", 'ROLE1!Four!first', true],
+            ["four", "first", new BindingRole("ROLE1", 2), ['album' => 1], "allowed", 'ROLE1!Four!first', true],
+            ["four", "first", new BindingRole("ROLE2", 1), ['album' => 1], "allowed", 'ROLE2!Four!first!1!1', true],
+            ["four", "first", new BindingRole("ROLE2", 2), ['album' => 1], null, 'ROLE2!Four!first!1!2', false],
+            ["four", "second", new BindingRole("ROLE1", 1), ['album' => 1], "allowed", 'ROLE1!Four!second', true],
+            ["four", "second", new BindingRole("ROLE1", 2), ['album' => 1], "allowed", 'ROLE1!Four!second', true],
+            ["four", "second", new BindingRole("ROLE2", 1), ['album' => 1], "allowed", 'ROLE2!Four!second', true],
+            ["four", "second", new BindingRole("ROLE2", 2), ['album' => 1], "allowed", 'ROLE2!Four!second', true],
+            ["three", "deny", new BindingRole("ROLE1", 2), ['album' => 1], "allowed", 'ROLE1!Three!*', true],
+            ["four", "first", new BindingRole("ROLE4", 1), ['album' => 1], "allowed", 'ROLE4!Four!*!1!1', true],
+            ["four", "first", new BindingRole("ROLE4", 2), ['album' => 1], null, 'ROLE4!Four!*!1!2', false],
+            ["three", "deny", new BindingRole("ROLE5", 2), ['album' => 1], null, 'ROLE5!*!*!2', false],
+            ["three", "deny", new BindingRole("ROLE5", 3), ['album' => 1], "allowed", 'ROLE5!*!*!3', true],
+        ];
     }
 }
